Improve serialization enough to make this landable
I haven't pursued mmap and such yet as it requires non-trivial
additional changes, and those can be made as part of a followup too.
