+++ runtime/near-vm-runner/Cargo.toml
@@ -26,18 +26,19 @@ once_cell.workspace = true
 parity-wasm = { workspace = true, optional = true }
 prefix-sum-vec.workspace = true
 ripemd.workspace = true
+rustix = { workspace = true, features = [ "fs" ] }
 serde_repr.workspace = true
 serde.workspace = true
 sha2.workspace = true
 sha3.workspace = true
 stdx.workspace = true
 strum.workspace = true
+tempfile.workspace = true
 thiserror.workspace = true
 tracing.workspace = true
 wasm-encoder = { workspace = true, optional = true }
 wasmparser = { workspace = true, optional = true }
 wasmtime = { workspace = true, optional = true }
-rustix = { workspace = true, features = [ "fs" ] }
 
 near-crypto.workspace = true
 near-primitives-core.workspace = true
@@ -72,8 +73,8 @@ hex.workspace = true
 near-test-contracts.workspace = true
 rand.workspace = true
 serde_json = { workspace = true, features = ["preserve_order"] }
-wasm-smith.workspace = true
 wasmprinter.workspace = true
+wasm-smith.workspace = true
 wat.workspace = true
 
 [features]
