+++ runtime/near-vm-runner/src/cache.rs
@@ -8,6 +8,8 @@ use near_parameters::vm::VMKind;
 use near_primitives_core::hash::CryptoHash;
 use std::collections::HashMap;
 use std::fmt;
+use std::io::{Read, Write};
+use std::path::PathBuf;
 use std::sync::{Arc, Mutex};
 
 #[derive(Debug, Clone, BorshSerialize)]
@@ -82,6 +84,95 @@ impl fmt::Debug for MockCompiledContractCache {
         let guard = self.store.lock().unwrap();
         let hm: &HashMap<_, _> = &*guard;
         fmt::Debug::fmt(hm, f)
+    }
+}
+
+pub struct FilesystemCompiledContractCache {
+    dir: rustix::fd::OwnedFd,
+}
+
+impl FilesystemCompiledContractCache {
+    pub fn new() -> Self {
+        let path = std::env::var_os("NEAR_CONTRACT_CACHE");
+        let path = path.map(PathBuf::from).unwrap_or_else(|| {
+            PathBuf::from(std::env::var_os("HOME").expect("$HOME"))
+                .join(".near")
+                .join("data")
+                .join("contracts")
+        });
+        std::fs::create_dir_all(&path).expect("create contract directory");
+        let dir = rustix::fs::open(path, rustix::fs::OFlags::DIRECTORY, rustix::fs::Mode::empty())
+            .expect("open contract dir");
+        Self { dir }
+    }
+}
+
+const ERROR_TAG: u8 = b'\n';
+const CODE_TAG: u8 = 0b10010101;
+/// Cache for compiled contracts code in plain filesystem.
+impl CompiledContractCache for FilesystemCompiledContractCache {
+    fn put(&self, key: &CryptoHash, value: CompiledContract) -> std::io::Result<()> {
+        use rustix::fs::{Mode, OFlags};
+        let final_filename = key.to_string();
+        let mut temp_file = tempfile::Builder::new().make_in("", |filename| {
+            let mode = Mode::RUSR | Mode::WUSR | Mode::RGRP | Mode::WGRP;
+            let flags = OFlags::CREATE | OFlags::TRUNC | OFlags::WRONLY;
+            Ok(std::fs::File::from(rustix::fs::openat(&self.dir, dbg!(filename), flags, mode)?))
+        })?;
+        // This section manually "serializes" the data. The cache is quite sensitive to
+        // unnecessary overheads and in order to enable things like mmap-based file access, we want
+        // to have full control of what has been written.
+        match value {
+            CompiledContract::CompileModuleError(e) => {
+                borsh::to_writer(&mut temp_file, &e)?;
+                temp_file.write_all(&[ERROR_TAG])?;
+            }
+            CompiledContract::Code(bytes) => {
+                temp_file.write_all(&bytes)?;
+                // Writing the tag at the end gives us well aligned buffer of the data above which
+                // is necessary for 0-copy deserialization later on.
+                temp_file.write_all(&[CODE_TAG])?;
+            }
+        }
+        let temp_filename = temp_file.into_temp_path();
+        // This is atomic, so there wouldn't be instances where getters see an intermediate state.
+        rustix::fs::renameat(&self.dir, &*temp_filename, &self.dir, final_filename)?;
+        // Don't attempt deleting the temporary file now that it has been moved.
+        std::mem::forget(temp_filename);
+        Ok(())
+    }
+
+    fn get(&self, key: &CryptoHash) -> std::io::Result<Option<CompiledContract>> {
+        use rustix::fs::{Mode, OFlags};
+        let filename = key.to_string();
+        let mode = Mode::empty();
+        let flags = OFlags::RDONLY;
+        let file = rustix::fs::openat(&self.dir, &filename, flags, mode);
+        let file = match file {
+            Err(rustix::io::Errno::NOENT) => return Ok(None),
+            Err(e) => return Err(e.into()),
+            Ok(file) => file,
+        };
+        let stat = rustix::fs::fstat(&file)?;
+        // TODO: explore mmaping the file and lending the map to the caller via a closure callback.
+        // This would require some additional refactor work, but would likely help us to reduce the
+        // system call overhead in this area.
+        let mut buffer = Vec::with_capacity(stat.st_size.try_into().unwrap());
+        let mut file = std::fs::File::from(file);
+        file.read_to_end(&mut buffer)?;
+        match buffer.pop() {
+            // The file turns out to be empty/truncated? Treat as if there's no cached file.
+            None => Ok(None),
+            Some(CODE_TAG) => Ok(Some(CompiledContract::Code(buffer))),
+            Some(ERROR_TAG) => {
+                Ok(Some(CompiledContract::CompileModuleError(borsh::from_slice(&buffer)?)))
+            }
+            // File is malformed? For this code, since we're talking about a cache lets just treat
+            // it as if there is no cached file as well. The cached file may eventually be
+            // overwritten with a valid copy. And since we can compile a new copy, there doesn't
+            // seem to be much reason to possibly crash the node due to this.
+            Some(_) => Ok(None),
+        }
     }
 }
