+++ runtime/near-vm-runner/src/lib.rs
@@ -27,10 +27,13 @@ mod wasmer_runner;
 mod wasmtime_runner;
 
 pub use crate::logic::with_ext_cost_counter;
-pub use cache::{get_contract_cache_key, precompile_contract, MockCompiledContractCache};
+pub use cache::{
+    get_contract_cache_key, precompile_contract, FilesystemCompiledContractCache,
+    MockCompiledContractCache,
+};
 pub use code::ContractCode;
 pub use profile::ProfileDataV3;
-pub use runner::{run, FilesystemCompiledContractCache, VM};
+pub use runner::{run, VM};
 
 /// This is public for internal experimentation use only, and should otherwise be considered an
 /// implementation detail of `near-vm-runner`.
